move blob serailization to separate cpp file
