+++ ydb/core/persqueue/ya.make
@@ -3,6 +3,7 @@ LIBRARY()
 SRCS(
     actor_persqueue_client_iface.h
     blob.cpp
+    blob_serialization.cpp
     common_app.cpp
     cluster_tracker.cpp
     event_helpers.cpp
